Split API into different files
