+++ api/api.php
@@ -2,15 +2,26 @@
 
 require_once __DIR__ . '/api-base.php';
 require_once __DIR__ . '/api-search.php';
+require_once __DIR__ . '/api-replace.php';
+require_once __DIR__ . '/api-source.php';
 require_once __DIR__ . '/api-settings.php';
 require_once __DIR__ . '/api-plugin.php';
 
 define( 'SEARCHREGEX_API_NAMESPACE', 'search-regex/v1' );
 
 class Search_Regex_Api {
-	/** @var Search_Regex_Api|null */
+	/**
+	 * Instance variable
+	 *
+	 * @var Search_Regex_Api|null
+	 **/
 	private static $instance = null;
-	/** @var Array */
+
+	/**
+	 * Array of endpoint routes
+	 *
+	 * @var Array
+	 **/
 	private $routes = array();
 
 	/**
@@ -32,6 +43,8 @@ class Search_Regex_Api {
 		$wpdb->hide_errors();
 
 		$this->routes[] = new Search_Regex_Api_Search( SEARCHREGEX_API_NAMESPACE );
+		$this->routes[] = new Search_Regex_Api_Replace( SEARCHREGEX_API_NAMESPACE );
+		$this->routes[] = new Search_Regex_Api_Source( SEARCHREGEX_API_NAMESPACE );
 		$this->routes[] = new Search_Regex_Api_Plugin( SEARCHREGEX_API_NAMESPACE );
 		$this->routes[] = new Search_Regex_Api_Settings( SEARCHREGEX_API_NAMESPACE );
 	}
